yadisk/requirements.txt
python-telegram-bot==20.3
httpx==0.27.0
pydantic==2.8.2
tomli==2.0.1